🌱 [scanner] fix: split analytics-events.test.ts (1546 lines) into focused modules
Split the oversized analytics-events.test.ts test file into 3 focused modules:
- analytics-events.cards-interactions.test.ts (684 lines) - Dashboard, cards, search, and UI interactions
- analytics-events.system-operations.test.ts (639 lines) - Auth, settings, APIs, and cluster operations
- analytics-events.engagement-lifecycle.test.ts (668 lines) - Nudges, recommendations, features, and ACMM

This addresses issue #18598 by reducing one of the largest oversized test files.
Each new file contains logically grouped test suites with proper structure and closure.
